move cspdarknet.py to yolof_base
